requirements.txt
pyTelegramBotAPI
pytest
jproperties
flake8
matplotlib
coverage
pytest-mock
email-validator==1.3.0
mock
coverage==6.5.0
configparser
pandas
numpy
telebot-calendar==1.2